+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/dto/RuleNameDto.java
@@ -6,6 +6,7 @@ import lombok.RequiredArgsConstructor;
 
 import javax.validation.constraints.NotBlank;
 
+
 @Data
 @AllArgsConstructor
 @RequiredArgsConstructor
