+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/service/impl/CurvePointServiceImpl.java
@@ -1,10 +1,10 @@
 package com.nnk.springboot.service.impl;
 
-import com.nnk.springboot.NotFoundException;
 import com.nnk.springboot.domain.CurvePoint;
 import com.nnk.springboot.dto.CurvePointDto;
 import com.nnk.springboot.repositories.CurvePointRepository;
 import com.nnk.springboot.service.CurvePointService;
+import com.nnk.springboot.util.exceptions.NotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
