+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/service/impl/TradeServiceImpl.java
@@ -1,10 +1,10 @@
 package com.nnk.springboot.service.impl;
 
-import com.nnk.springboot.NotFoundException;
 import com.nnk.springboot.domain.Trade;
 import com.nnk.springboot.dto.TradeDto;
 import com.nnk.springboot.repositories.TradeRepository;
 import com.nnk.springboot.service.TradeService;
+import com.nnk.springboot.util.exceptions.NotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
