+++ Poseiden-skeleton/src/main/java/com/nnk/springboot/util/exceptions/NotFoundException.java
@@ -1,4 +1,4 @@
-package com.nnk.springboot;
+package com.nnk.springboot.util.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
